Moved *Transaction structs from account to transaction.rs
