+++ workspaces/src/network/account.rs
@@ -1,11 +1,7 @@
-use std::convert::TryInto;
-
-use near_crypto::KeyType;
-
-use crate::types::{AccountId, Balance, InMemorySigner, SecretKey};
+use crate::types::{AccountId, Balance, InMemorySigner};
 use crate::{Network, Worker};
 
-use super::transaction::{CallArgs, Transaction};
+use super::transaction::{CallTransaction, CreateAccountTransaction, Transaction};
 use super::{CallExecution, CallExecutionDetails, ViewResultDetails};
 
 pub struct Account {
@@ -194,141 +190,5 @@ impl Contract {
     /// to send the batched transaction to the network.
     pub fn batch<'a, T: Network>(&self, worker: &'a Worker<T>) -> Transaction<'a> {
         Transaction::new(worker.client(), self.signer().clone(), self.id().clone())
-    }
-}
-
-pub struct CallTransaction<'a, T> {
-    worker: &'a Worker<T>,
-    signer: InMemorySigner,
-    contract_id: AccountId,
-    call_args: CallArgs,
-}
-
-impl<'a, T: Network> CallTransaction<'a, T> {
-    fn new(
-        worker: &'a Worker<T>,
-        contract_id: AccountId,
-        signer: InMemorySigner,
-        function: &str,
-    ) -> Self {
-        Self {
-            worker,
-            signer,
-            contract_id,
-            call_args: CallArgs::new(function),
-        }
-    }
-
-    pub fn args(mut self, args: Vec<u8>) -> Self {
-        self.call_args = self.call_args.args(args);
-        self
-    }
-
-    pub fn args_json<U: serde::Serialize>(mut self, args: U) -> anyhow::Result<Self> {
-        self.call_args = self.call_args.args_json(args)?;
-        Ok(self)
-    }
-
-    pub fn args_borsh<U: borsh::BorshSerialize>(mut self, args: U) -> anyhow::Result<Self> {
-        self.call_args = self.call_args.args_borsh(args)?;
-        Ok(self)
-    }
-
-    pub fn deposit(mut self, deposit: u128) -> Self {
-        self.call_args = self.call_args.deposit(deposit);
-        self
-    }
-
-    pub fn gas(mut self, gas: u64) -> Self {
-        self.call_args = self.call_args.gas(gas);
-        self
-    }
-
-    pub async fn transact(self) -> anyhow::Result<CallExecutionDetails> {
-        self.worker
-            .client()
-            .call(
-                &self.signer,
-                &self.contract_id,
-                self.call_args.function,
-                self.call_args.args,
-                self.call_args.gas,
-                self.call_args.deposit,
-            )
-            .await
-            .map(Into::into)
-    }
-
-    pub async fn view(self) -> anyhow::Result<ViewResultDetails> {
-        self.worker
-            .client()
-            .view(
-                self.contract_id,
-                self.call_args.function,
-                self.call_args.args,
-            )
-            .await
-    }
-}
-
-pub struct CreateAccountTransaction<'a, 'b, T> {
-    worker: &'a Worker<T>,
-    signer: InMemorySigner,
-    parent_id: AccountId,
-    new_account_id: &'b str,
-
-    initial_balance: Balance,
-    secret_key: Option<SecretKey>,
-}
-
-impl<'a, 'b, T> CreateAccountTransaction<'a, 'b, T>
-where
-    T: Network,
-{
-    fn new(
-        worker: &'a Worker<T>,
-        signer: InMemorySigner,
-        parent_id: AccountId,
-        new_account_id: &'b str,
-    ) -> Self {
-        Self {
-            worker,
-            signer,
-            parent_id,
-            new_account_id,
-            initial_balance: 100000000000000000000000,
-            secret_key: None,
-        }
-    }
-
-    pub fn initial_balance(mut self, initial_balance: Balance) -> Self {
-        self.initial_balance = initial_balance;
-        self
-    }
-
-    pub fn keys(mut self, secret_key: SecretKey) -> Self {
-        self.secret_key = Some(secret_key);
-        self
-    }
-
-    pub async fn transact(self) -> anyhow::Result<CallExecution<Account>> {
-        let sk = self
-            .secret_key
-            .unwrap_or_else(|| SecretKey::from_seed(KeyType::ED25519, "subaccount.seed"));
-        let id: AccountId = format!("{}.{}", self.new_account_id, self.parent_id).try_into()?;
-
-        let outcome = self
-            .worker
-            .client()
-            .create_account(&self.signer, &id, sk.public_key(), self.initial_balance)
-            .await?;
-
-        let signer = InMemorySigner::from_secret_key(id.clone(), sk);
-        let account = Account::new(id, signer);
-
-        Ok(CallExecution {
-            result: account,
-            details: outcome.into(),
-        })
     }
 }
